+++ wp-content/themes/labor-day/inc/plugins/acf/acf-classes/class-image.php
